Fix building with clang -Weverything
- iffe.sh: Look for 'error:' rather than 'error' when determining
  the outcome of a compile. This fixes a bug that caused iffe to
  incorrectly mark a compiler warning for the 'ferror' function
  as a compiler error. (This consequently fixes multiple io.sh
  test failures that occured when compiling with clang's -Weverything.
  It also fixes the iffe test for rlim_t.)
- Rename the _ast_signal() function to ast_signal, and make its
  uses explicit to avoid confusion with the separate signal(2).
  This is better for code clarity and fixes a compiler error that
  occurs when attempting to build with Clang 20.1.7 while using
  -Weverything.
- Fix many -Wunused warnings by adding missing NOT_USED() usages for
  unused parameters, deleting unused macros, and moving misplaced macro
  definitions.
- Fix a few lingering sign conversion warnings too.
- Move the SOCK_CLOEXEC fallback definition to avoid possible compiler
  errors. (The strace output looks fine, thankfully.)
- sfdcdio.c: Amend straggler code to avoid null pointer arithmetic.
- Delete the unused pathcheck() function (which suffered from some
  bit rot).

+++ src/cmd/ksh93/include/path.h
@@ -68,7 +68,7 @@ typedef struct pathcomp
 extern void		path_newdir(Pathcomp_t*);
 extern Pathcomp_t	*path_dirfind(Pathcomp_t*,const char*,int);
 extern Pathcomp_t	*path_unsetfpath(void);
-extern Pathcomp_t	*path_addpath(Pathcomp_t*,const char*,int);
+extern Pathcomp_t	*path_addpath(Pathcomp_t*,const char*,uint16_t);
 extern Pathcomp_t	*path_dup(Pathcomp_t*);
 extern void		path_delete(Pathcomp_t*);
 extern void 		path_settrackedalias(const char*,Pathcomp_t*);
